+++ lib/ui/wellness/todo/WellnessCreateToDoItemPanel.dart
@@ -343,6 +343,7 @@ class _WellnessCreateToDoItemPanelState extends State<WellnessCreateToDoItemPane
   }
 
   void _onTapCurrentCategory() {
+    _hideKeyboard();
     _categoriesDropDownVisible = !_categoriesDropDownVisible;
     if (mounted) {
       setState(() {});
@@ -350,6 +351,7 @@ class _WellnessCreateToDoItemPanelState extends State<WellnessCreateToDoItemPane
   }
 
   void _onTapCategory(ToDoCategory? category) {
+    _hideKeyboard();
     if (_category != category) {
       _category = category;
     }